# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Fresh Platform API
  description: |
    Complete API documentation for the Fresh Platform - a Georgian AI & Technology platform.
    
    ## Authentication
    Most endpoints require JWT authentication via Bearer token in the Authorization header.
    Admin endpoints require a separate admin token via Cookie (`admin_session`).
    
    ## Rate Limiting
    - Search API: 30 requests per minute per IP
    - Admin login: Brute-force protection with lockout
    
    ## Georgian Language
    Many error messages and content are in Georgian.
    
  version: 1.0.0
  contact:
    name: Andrew Altair
    email: andrewaltair@icloud.com
  license:
    name: MIT
servers:
  - url: https://api.example.com
    description: Production server
  - url: http://localhost:3000
    description: Development server
...
paths:
...
  /api/posts/parse-ai:
    post:
      tags: [Posts]
      summary: Parse content with AI
      description: Structure raw content into sections
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [rawContent]
              properties:
                rawContent:
                  type: string
      responses:
        '200':
          description: Parsed content
          content:
            application/json:
              schema:
                type: object
                properties:
                  title:
                    type: string
                  excerpt:
                    type: string
                  sections:
                    type: array
                    items:
                      type: object
                      properties:
                        icon:
                          type: string
                        title:
                          type: string
                        content:
                          type: string
                        type:
                          type: string
                  tags:
                    type: array
                    items:
                      type: string
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    adminCookie:
      type: apiKey
      in: cookie
      name: admin_session
...